refactor: User 엔티티와 리포지토리의 AuthProvider import를 수정했습니다
